+++ front/components/SongRow.tsx
@@ -3,32 +3,13 @@ import Song, { SongWithArtist } from '../models/Song';
 import RowCustom from './RowCustom';
 import TextButton from './TextButton';
 import { MaterialIcons } from '@expo/vector-icons';
+import DurationComponent from './DurationComponent';
 
 type SongRowProps = {
 	song: Song | SongWithArtist; // TODO: remove Song
 	isLiked: boolean;
 	onPress: () => void;
 	handleLike: (state: boolean, songId: number) => Promise<void>;
-};
-
-type DurationInfoProps = {
-	length: number;
-};
-
-const DurationInfo = ({ length }: DurationInfoProps) => {
-	const minutes = Math.floor(length / 60);
-	const seconds = Math.round(length - minutes * 60);
-
-	return (
-		<Text
-			style={{
-				flexShrink: 0,
-			}}
-			fontSize={'sm'}
-		>
-			{`${minutes}'${seconds}` ?? '--.--'}
-		</Text>
-	);
 };
 
 const SongRow = ({ song, onPress, handleLike, isLiked }: SongRowProps) => {
@@ -75,7 +56,8 @@ const SongRow = ({ song, onPress, handleLike, isLiked }: SongRowProps) => {
 					>
 						{song.artistId ?? 'artist'}
 					</Text>
-					<DurationInfo length={song.details.length} />
+					{/* <DurationInfo length={song.details.length} /> */}
+					<DurationComponent length={song.details.length} />
 				</HStack>
 				<IconButton
 					colorScheme="rose"
